fix `-Werror=arith-conversion`
and indentation

Ref: https://github.com/curl/curl/actions/runs/7148682295/job/19469980314?pr=12492#step:29:1235

```
inet_pton.c:115:35: error: conversion to 'unsigned int' from 'int' may change the sign of the result [-Werror=arith-conversion]
  115 |       unsigned int val = *tp * 10 + (unsigned int)(pch - digits);
      |                                   ^
```
Ref: https://dev.azure.com/daniel0244/curl/_build/results?buildId=18964&view=logs&jobId=fd8b539a-45be-52ee-427b-3ad3fdec5ebd&j=81bb90a1-709c-5261-a92d-4869708fd1b6&t=5cd3fffe-65e4-58bd-82d3-f0b036b20e6c

+++ lib/doh.c
@@ -444,7 +444,7 @@ static DOHcode skipqname(const unsigned char *doh, size_t dohlen,
       return DOH_DNS_BAD_LABEL;
     if(dohlen < (*indexp + 1 + length))
       return DOH_DNS_OUT_OF_RANGE;
-    *indexp += 1 + length;
+    *indexp += (unsigned int)(1 + length);
   } while(length);
   return DOH_OK;
 }
@@ -456,14 +456,15 @@ static unsigned short get16bit(const unsigned char *doh, int index)
 
 static unsigned int get32bit(const unsigned char *doh, int index)
 {
-   /* make clang and gcc optimize this to bswap by incrementing
-      the pointer first. */
-   doh += index;
+  /* make clang and gcc optimize this to bswap by incrementing
+     the pointer first. */
+  doh += index;
 
-   /* avoid undefined behavior by casting to unsigned before shifting
-      24 bits, possibly into the sign bit. codegen is same, but
-      ub sanitizer won't be upset */
-  return ( (unsigned)doh[0] << 24) | (doh[1] << 16) |(doh[2] << 8) | doh[3];
+  /* avoid undefined behavior by casting to unsigned before shifting
+     24 bits, possibly into the sign bit. codegen is same, but
+     ub sanitizer won't be upset */
+  return ((unsigned)doh[0] << 24) | ((unsigned)doh[1] << 16) |
+         ((unsigned)doh[2] << 8) | doh[3];
 }
 
 static DOHcode store_a(const unsigned char *doh, int index, struct dohentry *d)
